requirements.txt
Django==4.2.4; python_version >= '3.11'
psycopg2-binary==2.9.6
google-cloud-secret-manager==2.16.1
argon2-cffi==23.1.0
snoop==0.4.3
django-structlog==5.3.0
rich==13.5.2
django-extensions==3.2.3
django-allauth==0.54.0
django-htmx==1.17.2
django-silk==5.0.3
django-browser-reload==1.11.0
django-debug-toolbar==4.1.0
django-environ==0.11.2
pyyaml==6.0.1
jwt==1.3.1
pytest-factoryboy==2.5.1
pytest-django==4.7.0
django-environ==0.11.2
pytest-playwright==0.4.4
pycmarkgfm==1.2.1
markdown2==2.4.13
bleach==6.2.0